+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/OpANNN.java
@@ -1,7 +1,7 @@
 package dev.krk.emulator.chipsekiz.opcodes;
 
 public class OpANNN extends OpPNNN {
-    public OpANNN(int address) {
+    public OpANNN(short address) {
         super(0xA, address);
     }
 }
