+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/OpBNNN.java
@@ -1,7 +1,7 @@
 package dev.krk.emulator.chipsekiz.opcodes;
 
 public class OpBNNN extends OpPNNN {
-    public OpBNNN(int address) {
+    public OpBNNN(short address) {
         super(0xB, address);
     }
 }
